requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-gis==1.0
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
